+++ Editor/GitHubDownloadWindow.cs
@@ -1,42 +1,43 @@
-﻿using System;
+﻿
+using System;
+using System.Collections.Generic;
 using System.IO;
-
+using System.Linq;
 using System.Net;
+using System.Text.RegularExpressions;
+using System.Threading.Tasks;
 using UnityEditor;
 using UnityEngine;
-using System.Linq;
-using System.Text.RegularExpressions;
-using UnityEditorInternal;
-using System.Threading.Tasks;
 
 using static System.IO.Path;
+
 using E = Hananoki.GitHubDownload.GitHubDownloadSettingsEditor;
 
 namespace Hananoki.GitHubDownload {
 
 	public class GitHubDownloadWindow : EditorWindow {
 
+		public static GitHubDownloadWindow s_window;
+
 		string githubURL = "https://api.github.com/repos";
 
-		//string[] urls = {
-		//	"https://github.com/hananoki/HananokiSharedModule.git",
-		//	"https://github.com/hananoki/HierarchyDropDown.git",
-		//	};
-
 		public int m_selectURL;
-		public ReleaseJson m_js;
+		public List<ReleaseJson> m_js;
 
 		public class Styles {
 			public GUIStyle ExposablePopupItem;
 			public GUIStyle boldLabel;
 			public GUIStyle miniBoldLabel;
 			public GUIStyle helpBox;
+			public GUIStyle toolbarDropDown;
 
 			public Texture2D Icon;
 			public Texture2D IconSetting;
 			public Texture2D IconError;
 			public Texture2D IconInfo;
 			public Texture2D[] IconWaitSpin;
+			public Texture2D IconRefresh;
+			public Texture2D IconSceneAsset;
 			public Styles() {
 				ExposablePopupItem = new GUIStyle( "ExposablePopupItem" );
 				ExposablePopupItem.margin = new RectOffset( 2, 2, 2, 2 );
@@ -52,35 +53,27 @@ namespace Hananoki.GitHubDownload {
 				Icon = EditorGUIUtility.FindTexture( "winbtn_mac_inact" );
 				IconWaitSpin = Resources.FindObjectsOfTypeAll<Texture2D>().Where( x => x.name.Contains( "WaitSpin" )).OrderBy( x => x.name ).ToArray();
 				IconSetting = EditorGUIUtility.FindTexture( "SettingsIcon" );
+				IconRefresh = EditorGUIUtility.FindTexture( "Refresh" );
+				IconSceneAsset = Resources.FindObjectsOfTypeAll<Texture2D>().Where( x => x.name == "SceneAsset Icon" ).ToArray()[0];
+
+				toolbarDropDown = new GUIStyle( EditorStyles.toolbarDropDown );
+				toolbarDropDown.alignment = TextAnchor.MiddleCenter;
+				toolbarDropDown.padding.right = 12;
 			}
 		}
 
 		public static Styles s_styles;
 
 
-		public string[] m_downloadFiles;
+		public List<string> m_downloadDirs;
 		static Vector2 m_scroll;
 
 		bool indexChanged;
-		bool networking;
+
+		bool enableLatest;
+
+		string[] showRelease = { "Releases", "Latest Releases" };
 		
-		public string networkingMsg;
-
-		bool networkError;
-		public string networkingErrorMsg;
-
-		public static string gitHubCacheDirectory {
-			get {
-#if UNITY_EDITOR_WIN
-				return GetDirectoryName( ( GetDirectoryName( InternalEditorUtility.unityPreferencesFolder ) ) ).Replace( '\\', '/' ) + "/GitHub";
-#elif UNITY_EDITOR_OSX
-			return InternalEditorUtility.unityPreferencesFolder + "/" + "../../../Unity/GitHub";
-#else
-			return "";
-#endif
-			}
-		}
-
 
 
 		[MenuItem( "Window/GitHubDownload" )]
@@ -89,38 +82,23 @@ namespace Hananoki.GitHubDownload {
 			window.titleContent = new GUIContent( "GitHubDownload" );
 		}
 
+		new public static void Repaint() {
+			((EditorWindow) s_window )?.Repaint();
+		}
 
 		void OnEnable() {
+			s_window = this;
 			E.Load();
 			MakeDownloadList();
-			ReadJson();
+			ReadWebResponseToFile( enableLatest );
 			indexChanged = true;
-
-			
 		}
 
-		static float curTime;
-		static float lastTime;
-		public static int m_count;
-		static float m_watiTime;
 
-		public  void updateThreadSync() {
-			curTime = Time.realtimeSinceStartup;
-			float deltaTime = (float) ( curTime - lastTime );
-			lastTime = curTime;
-
-			m_watiTime -= deltaTime;
-			if( m_watiTime < 0 ) {
-				m_watiTime = 0.1250f;
-
-				m_count++;
-				if( 12 <= m_count ) {
-					m_count = 0;
-				}
-				Repaint();
-			}
+		string MakeOutputPath( string gitURL ) {
+			var info = ParseURL( GetCurrentURL() );
+			return $"{E.gitHubCacheDirectory}/{info[ 0 ]}/{info[ 1 ]}";
 		}
-
 
 		string[] ParseURL( string gitURL ) {
 			if( string.IsNullOrEmpty( gitURL ) ) return null;
@@ -131,122 +109,181 @@ namespace Hananoki.GitHubDownload {
 		}
 
 
-
-		string GetCurrentURL() {
-			if( E.i.urls.Count == 0 ) return string.Empty;
+		bool AdjustSelectURL() {
+			if( E.i.urls.Count == 0 ) return false;
 			if( m_selectURL < 0 ) {
 				m_selectURL = 0;
 				indexChanged = true;
+				return true;
 			}
-			if( E.i.urls.Count <= m_selectURL  ) {
-				m_selectURL = E.i.urls.Count -1;
+			if( E.i.urls.Count <= m_selectURL ) {
+				m_selectURL = E.i.urls.Count - 1;
 				indexChanged = true;
+				return true;
 			}
+			return false;
+		}
+
+
+		string GetCurrentURL() {
+			if( E.i.urls.Count == 0 ) return string.Empty;
+
+			AdjustSelectURL();
+
 			return E.i.urls[ m_selectURL ];
 		}
 
 
 		void MakeDownloadList() {
-			m_downloadFiles = new string[ 0 ];
+			m_downloadDirs = new List<string>();
 
 			var info = ParseURL( GetCurrentURL() );
-			var outputDirectory = $"{gitHubCacheDirectory}/{info[ 0 ]}/{info[ 1 ]}";
+			if( info == null ) return;
+			var outputDirectory = $"{E.gitHubCacheDirectory}/{info[ 0 ]}/{info[ 1 ]}";
 
 			if( !Directory.Exists( outputDirectory ) ) return;
 
-			m_downloadFiles = Directory.GetDirectories( outputDirectory );
+			m_downloadDirs = Directory.GetDirectories( outputDirectory ).ToList();
 		}
 
 
-		void SaveJson( ReleaseJson js ) {
-			if( js == null ) return;
-			var info = ParseURL( E.i.urls[ m_selectURL ] );
-			var opath = $"{gitHubCacheDirectory}/{info[ 0 ]}/{info[ 1 ]}";
+		void WriteWebResponseToFile( string content, bool latest ) {
+			var opath = MakeOutputPath( GetCurrentURL() );
 
 			if( !Directory.Exists( opath ) ) {
 				Directory.CreateDirectory( opath );
 			}
 
-			using( var st = new StreamWriter( $"{opath}/releases_latest.json" ) ) {
-				st.Write( JsonUtility.ToJson( js ) );
+			if( latest ) {
+				opath = $"{opath}/releases_latest.json";
+			}
+			else {
+				opath = $"{opath}/releases.json";
+			}
+
+			
+
+			using( var st = new StreamWriter( opath ) ) {
+				st.Write(  content  );
 			}
 		}
 
 
-		void ReadJson() {
-			m_js = null;
-			var info = ParseURL( E.i.urls[ m_selectURL ] );
-			var opath = $"{gitHubCacheDirectory}/{info[ 0 ]}/{info[ 1 ]}";
+		void ReadWebResponseToFile( bool latest ) {
+			m_js = new List<ReleaseJson>();
+			var info = ParseURL( GetCurrentURL() );
+			if( info == null ) return;
+
+			var opath = $"{E.gitHubCacheDirectory}/{info[ 0 ]}/{info[ 1 ]}";
 
 			if( !Directory.Exists( opath ) ) return;
 
-			using( var st = new StreamReader( $"{opath}/releases_latest.json" ) ) {
-				m_js = JsonUtility.FromJson<ReleaseJson>( st.ReadToEnd() );
+			if( latest ) {
+				var fname = $"{opath}/releases_latest.json";
+				if( !File.Exists( fname ) ) return;
+
+				using( var st = new StreamReader( fname ) ) {
+					var jss = JsonUtility.FromJson<ReleaseJson>( st.ReadToEnd() );
+					m_js.Add( jss );
+				}
+			}
+			else {
+				var fname = $"{opath}/releases.json";
+				if( !File.Exists( fname ) ) return;
+
+				using( var st = new StreamReader( fname ) ) {
+					var content = st.ReadToEnd();
+					var content2 = "{\"Items\":" + content + "}";
+					var jss = JsonHelper.FromJson<ReleaseJson>( content2 );
+					m_js.AddRange( jss );
+				}
 			}
 		}
 
 
-		async void GetReleasesLatest( string name, string repoName ) {
+		async void GetReleasesResponse( string name, string repoName, bool latest ) {
 			string content = string.Empty;
-			m_js = null;
-			networking = true;
-			networkError = false;
-			networkingMsg = "Download Release Latest Json";
-			EditorApplication.update += updateThreadSync;
+			m_js = new List<ReleaseJson>();
+
 			await Task.Run( () => {
 				try {
+					using( new RequestStatusScope(  latest ? "Downloading Release Latest ..." : "Downloading Release ..." ) )
 					using( var client = new WebClient() ) {
-						var url = $"{githubURL}/{name}/{repoName}/releases/latest";
-
+						var url = $"{githubURL}/{name}/{repoName}/releases";
+						if( latest ) {
+							url += "/latest";
+						}
 						client.Headers.Add( "User-Agent", "Nothing" );
 						//System.Threading.Thread.Sleep( 10000 );
 						content = client.DownloadString( url );
 					}
+
+					if( string.IsNullOrEmpty( content ) ) return;
+
+					
+					if( latest ) {
+						var rj = JsonUtility.FromJson<ReleaseJson>( content );
+						if( rj == null ) return;
+						m_js.Add( rj );
+					}
+					else {
+						var content2 = "{\"Items\":" + content + "}";
+						var jss = JsonHelper.FromJson<ReleaseJson>( content2 );
+						m_js.AddRange( jss );
+					}
+					WriteWebResponseToFile( content, latest );
+					
+
 				}
 				catch( Exception e ) {
 					Debug.LogException( e );
-					networkError = true;
-					networkingErrorMsg = e.Message;
+					RequestStatus.SetError( e );
 				}
 			} );
-			networking = false;
-			EditorApplication.update -= updateThreadSync;
-
-			if( string.IsNullOrEmpty( content ) ) return;
-
-			m_js = JsonUtility.FromJson<ReleaseJson>( content );
-			if( m_js == null ) return;
-			SaveJson( m_js );
+			Repaint();
 		}
 
-		void GetReleasesLatest( string gitURL ) {
+
+		void GetReleasesResponse( string gitURL, bool latest ) {
 			var m = ParseURL( gitURL );
-			GetReleasesLatest( m[ 0 ], m[ 1 ] );
+			GetReleasesResponse( m[ 0 ], m[ 1 ], latest );
 		}
 
 
-		public async void DownloadFile( string url, string name, string repoName, string tag ) {
-			var outputDirectory = $"{gitHubCacheDirectory}/{name}/{repoName}/{tag}";
+		public async void DownloadFile( string url, string name, string repoName, string tag, string extention = "" ) {
+			var outputDirectory = $"{E.gitHubCacheDirectory}/{name}/{repoName}/{tag}";
 
 			if( !Directory.Exists( outputDirectory ) ) {
 				Directory.CreateDirectory( outputDirectory );
 			}
 
-			networking = true;
-			networkingMsg = "Download File";
-			EditorApplication.update += updateThreadSync;
 			await Task.Run( () => {
-				var fname = GetFileName( url );
-				//System.Threading.Thread.Sleep( 10000 );
-				using( WebClient wc = new WebClient() ) {
-					wc.DownloadFile( new Uri( url ), outputDirectory + "/" + fname );
+				try {
+					string fname;
+					if( string.IsNullOrEmpty( extention ) ) {
+						fname = GetFileName( url );
+					}
+					else {
+						fname = $"{repoName}-{tag}{extention}";
+					}
+					//System.Threading.Thread.Sleep( 10000 );
+					using( new RequestStatusScope( "Download File " + GetFileName( url ) ) )
+					using( WebClient wc = new WebClient() ) {
+						wc.Headers.Add( "User-Agent", "Nothing" );
+						wc.DownloadFile( new Uri( url ), outputDirectory + "/" + fname );
+					}
+				}
+				catch( Exception e ) {
+					Debug.LogException( e );
+					RequestStatus.SetError( e );
 				}
 			} );
-			networking = false;
-			EditorApplication.update -= updateThreadSync;
 
 			MakeDownloadList();
+			Repaint();
 		}
+
+
 
 
 		void DrawToolbar() {
@@ -255,21 +292,24 @@ namespace Hananoki.GitHubDownload {
 				EditorApplication.ExecuteMenuItem( "Edit/Preferences..." );
 			}
 			EditorGUI.BeginChangeCheck();
+			bool force = false;
 			if( 0 < E.i.urls.Count ) {
-				m_selectURL = EditorGUILayout.Popup( m_selectURL, E.i.urls.Select( x => GetFileNameWithoutExtension( x ) ).ToArray(), EditorStyles.toolbarDropDown );
+				AdjustSelectURL();
+				m_selectURL = EditorGUILayout.Popup( m_selectURL, E.i.urls.Select( x => GetFileNameWithoutExtension( x ) ).ToArray(), s_styles.toolbarDropDown );
 			}
-
-			if( EditorGUI.EndChangeCheck()  ) {
+			E.i.showMode = EditorGUILayout.Popup( E.i.showMode, showRelease, s_styles.toolbarDropDown, GUILayout.Width( 120) );
+			if( EditorGUI.EndChangeCheck() || force ) {
+				enableLatest = E.i.showMode != 0;
 				MakeDownloadList();
-				ReadJson();
+				ReadWebResponseToFile( enableLatest );
 				indexChanged = true;
-				networkError = false;
+				RequestStatus.Reset();
 			}
-			GUILayout.FlexibleSpace();
+			if( GUILayout.Button( s_styles.IconRefresh, EditorStyles.toolbarButton ) ) {
+				GetReleasesResponse( GetCurrentURL(), enableLatest );
+			}
 
-			if( GUILayout.Button( "Get Releases Latest", EditorStyles.toolbarButton ) ) {
-				GetReleasesLatest( GetCurrentURL() );
-			}
+			GUILayout.FlexibleSpace();
 
 			GUILayout.EndHorizontal();
 		}
@@ -281,7 +321,7 @@ namespace Hananoki.GitHubDownload {
 			if( s_styles == null ) {
 				s_styles = new Styles();
 				MakeDownloadList();
-				ReadJson();
+				ReadWebResponseToFile( enableLatest );
 			}
 
 			DrawToolbar();
@@ -292,78 +332,96 @@ namespace Hananoki.GitHubDownload {
 				EditorGUILayout.HelpBox( "Set URL from preferences", MessageType.Info );
 				return;
 			}
-			if( networkError  ) {
-				EditorGUILayout.HelpBox( networkingErrorMsg, MessageType.Error );
+			if( RequestStatus.networkError  ) {
+				EditorGUILayout.HelpBox( RequestStatus.networkingErrorMsg, MessageType.Error );
 				return;
 			}
 
 			using( new GUILayout.HorizontalScope() ) {
 				bool force = false;
 				if( indexChanged ) {
-					if( m_js == null ) {
-						indexChanged = false;
+					indexChanged = false;
+					if( m_js == null || m_js.Count==0 ) {
 						force = true;
 					}
 				}
 				if( /*GUILayout.Button( "Get Releases Latest" ) ||*/ force ) {
-					Debug.Log( "Get Releases Latest" );
+					//Debug.Log( "Get Releases Latest" );
 					
-					GetReleasesLatest( GetCurrentURL() );
+					GetReleasesResponse( GetCurrentURL(), enableLatest );
 				}
 			}
 
 			using( var sc = new GUILayout.ScrollViewScope( m_scroll ) ) {
-				m_scroll = sc.scrollPosition;
-
 				if( m_js != null ) {
-					GUILayout.Label( info[1], s_styles.boldLabel );
-					var rc = GUILayoutUtility.GetLastRect();
-					rc.x += s_styles.boldLabel.CalcSize( new GUIContent( info[1] ) ).x;
-					rc.x += 8;
-					GUI.Label( rc, m_js.tag_name, s_styles.miniBoldLabel );
+					m_scroll = sc.scrollPosition;
+					if( m_js.Count == 0 ) {
+						if( !RequestStatus.networking ) {
+							EditorGUILayout.HelpBox( "No release", MessageType.Warning );
+						}
+					}
+					else {
+						foreach( var p in m_js ) {
+							using( new GUILayout.VerticalScope( s_styles.helpBox ) ) {
+								using( new GUILayout.HorizontalScope() ) {
+									p.toggle = GUIHelper.Foldout( p.toggle, $"{p.tag_name}" );
 
-					EditorGUILayout.LabelField( m_js.body, s_styles.helpBox );
+									var dldirs = m_downloadDirs.Where( x => GetFileName( x ) == p.tag_name ).ToList();
+									foreach( var dr in dldirs ) {
+										foreach( var fname in Directory.GetFiles( dr ) ) {
+											if( GetExtension( fname ) == ".unitypackage" ) {
+												var cont = new GUIContent( s_styles.IconSceneAsset );
+												var rc = GUILayoutUtility.GetRect( cont, GUIHelper.Styles.iconButton );
+#if UNITY_2019_3_OR_NEWER
+												rc.y += 1;
+#endif
+												if( GUIHelper.IconButton( rc, s_styles.IconSceneAsset ) ) {
+													AssetDatabase.ImportPackage( fname, true );
+												}
+											}
+										}
+									}
+								}
 
-					//var outputDirectory = $"{gitHubCacheDirectory}/{info[ 0 ]}/{info[ 1 ]}/{m_js.tag_name}";
-					foreach( var asset in m_js.assets ) {
-						var fname = GetFileName( asset.browser_download_url );
-						if( GUILayout.Button( new GUIContent( fname, s_styles.Icon ), s_styles.ExposablePopupItem, GUILayout.ExpandWidth( false ) ) ) {
-							DownloadFile( asset.browser_download_url, info[ 0 ], info[ 1 ], m_js.tag_name );
+								if( !p.toggle ) continue;
+
+								EditorGUILayout.LabelField( p.body, s_styles.helpBox );
+
+								//var outputDirectory = $"{gitHubCacheDirectory}/{info[ 0 ]}/{info[ 1 ]}/{m_js.tag_name}";
+								foreach( var asset in p.assets ) {
+									var fname = GetFileName( asset.browser_download_url );
+									if( GUILayout.Button( new GUIContent( fname, s_styles.Icon ), s_styles.ExposablePopupItem, GUILayout.ExpandWidth( false ) ) ) {
+										DownloadFile( asset.browser_download_url, info[ 0 ], info[ 1 ], p.tag_name );
+									}
+								}
+
+								if( !string.IsNullOrEmpty( p.zipball_url ) ) {
+									if( GUILayout.Button( new GUIContent( "Source code (zip)", s_styles.Icon ), s_styles.ExposablePopupItem, GUILayout.ExpandWidth( false ) ) ) {
+										DownloadFile( p.zipball_url, info[ 0 ], info[ 1 ], p.tag_name, ".zip" );
+									}
+								}
+								if( !string.IsNullOrEmpty( p.tarball_url ) ) {
+									if( GUILayout.Button( new GUIContent( "Source code (tar.gz)", s_styles.Icon ), s_styles.ExposablePopupItem, GUILayout.ExpandWidth( false ) ) ) {
+										DownloadFile( p.tarball_url, info[ 0 ], info[ 1 ], p.tag_name, ".tar.gz" );
+									}
+								}
+							}
 						}
 					}
 
-					if( !string.IsNullOrEmpty( m_js.zipball_url ) ) {
-						if( GUILayout.Button( new GUIContent( "Source code (zip)", s_styles.Icon ), s_styles.ExposablePopupItem, GUILayout.ExpandWidth( false ) ) ) {
-						}
-					}
-					if( !string.IsNullOrEmpty( m_js.tarball_url ) ) {
-						if( GUILayout.Button( new GUIContent( "Source code (tar.gz)", s_styles.Icon ), s_styles.ExposablePopupItem, GUILayout.ExpandWidth( false ) ) ) {
-						}
-					}
-				}
-
-
-				foreach( var p in m_downloadFiles ) {
-					if( GUILayout.Button( GetFileName( p ) ) ) {
-						
-						AssetDatabase.ImportPackage( Directory.GetFiles( p )[ 0 ] , true);
-					}
 				}
 			}
-
 			
-			if( networking ) {
+			if( RequestStatus.networking ) {
 				var last = GUILayoutUtility.GetLastRect();
-				var y = last.y /*+ last.height - 20*/;
-				last.y = y;
 				last.height = 20;
-				//EditorGUI.DrawRect( last, new Color( 0, 0, 1, 0.5f ) );
-				var cont = new GUIContent( networkingMsg, s_styles.IconWaitSpin[ m_count ] );
+				
+				var cont = new GUIContent( RequestStatus.networkingMsg, s_styles.IconWaitSpin[ RequestStatus.m_count ] );
 				last.width = EditorStyles.label.CalcSize( cont ).x;
 				last.x += 4;
 				last.width += 4;
 				last.y += 4;
-				EditorGUI.DrawRect( last, new Color( 1, 1, 1, 0.5f ) );
+				EditorGUI.DrawRect( last, new Color( 1, 1, 1, 1.0f ) );
 				GUI.Label( last, cont );
 			}
 		}