+++ Editor/PackageConfig.cs
@@ -3,7 +3,7 @@ namespace Hananoki.GitHubDownload {
 	public static class PackageInfo {
 		public const string name = "GitHubDownload";
 		public const string editorPrefName = "Hananoki.GitHubDownload";
-		public const string version = "0.3.3-preview";
+		public const string version = "0.4.0-preview";
 		//public static string projectSettingsPath => $"{Environment.CurrentDirectory}/ProjectSettings/BuildManagerSettings.json";
 	}
 }